requirements.txt
streamlit
PyPDF2
python-docx
pdf2docx
deep-translator